Remove bug in time management in probe\timer, add worker example in examples/localhost.php.

+++ classes/probe/timer.php
@@ -27,6 +27,6 @@ final class timer extends probe\generic
 
 	private static function now()
 	{
-		return microtime(true) * 10000;
+		return (int) microtime(true) * 10000;
 	}
 }
